drive-organizer/requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
rich==13.7.1
click==8.1.7
python-dotenv==1.0.1
